+++ packages/ui-components/src/radio/radio-group.tsx
@@ -173,9 +173,10 @@ function RadioGroup(props: RadioGroupProps): JSX.Element {
     const onChangeValue = (event: React.FormEvent<HTMLInputElement>): void => {
         const target = event.target as HTMLInputElement;
         const chosenIndex = Number(target.value);
-        setCurrentSelected(chosenIndex);
+        // controlled mode
         if (props.value !== undefined) {
             props.onChange?.(props.items[chosenIndex], event);
+            // uncontrolled mode
         } else {
             setCurrentSelected(chosenIndex);
         }